formatted code

+++ plotly-app/pages/utils/utils.py
@@ -2,9 +2,10 @@ import pandas as pd
 import os
 
 # paths relative to THIS script
-FBK_FILE_PATH        = "../../../FBK data/data_fbk_from_db.csv"
-APPA_FILE_PATH       = "../../data/21_22_APPA.csv"
+FBK_FILE_PATH = "../../../FBK data/data_fbk_from_db.csv"
+APPA_FILE_PATH = "../../data/21_22_APPA.csv"
 PREDICTION_FILE_PATH = "../../data/appa1_predictions.csv"
+
 
 def filter_fbk_data(dataframe: pd.DataFrame) -> pd.DataFrame:
     """
@@ -16,35 +17,34 @@ def filter_fbk_data(dataframe: pd.DataFrame) -> pd.DataFrame:
     Returns:
         pd.DataFrame: the filtered dataframe
     """
-    dataframe = dataframe.drop([
-        "node_name",
-        "g",
-        "h",
-        "th",
-        "cfg",
-        "iaq",
-        "co2",
-        "voc",
-        "iac_comp"
-    ], axis=1)
-    dataframe["ts"]         = pd.to_datetime(dataframe["ts"])
+    dataframe = dataframe.drop(
+        ["node_name", "g", "h", "th", "cfg", "iaq", "co2", "voc", "iac_comp"], axis=1
+    )
+    dataframe["ts"] = pd.to_datetime(dataframe["ts"])
     # Set CEST Summer time zone
-    dataframe["ts"]         += pd.Timedelta(hours=2)
+    dataframe["ts"] += pd.Timedelta(hours=2)
     dataframe["signal_res"] = dataframe["signal_res"].astype(float)
     dataframe["heater_res"] = dataframe["heater_res"].astype(float)
-    dataframe["volt"]       = dataframe["volt"].astype(float)
-    dataframe["p"]          = dataframe["p"].astype(float)
-    dataframe["t"]          = dataframe["t"].astype(float)
-    dataframe["rh"]         = dataframe["rh"].astype(float)
+    dataframe["volt"] = dataframe["volt"].astype(float)
+    dataframe["p"] = dataframe["p"].astype(float)
+    dataframe["t"] = dataframe["t"].astype(float)
+    dataframe["rh"] = dataframe["rh"].astype(float)
 
-    dataframe['sensor_description'] = dataframe['sensor_description'].str.split(pat="_").str.get(0)
+    dataframe["sensor_description"] = (
+        dataframe["sensor_description"].str.split(pat="_").str.get(0)
+    )
 
-    dataframe.drop_duplicates(['sensor_description','ts'], inplace = True)
+    dataframe.drop_duplicates(["sensor_description", "ts"], inplace=True)
 
-    dataframe["node_description"] = dataframe["node_description"].str.replace("Appa 1 - ", "")
-    dataframe["node_description"] = dataframe["node_description"].str.replace("Appa 2 - ", "")
+    dataframe["node_description"] = dataframe["node_description"].str.replace(
+        "Appa 1 - ", ""
+    )
+    dataframe["node_description"] = dataframe["node_description"].str.replace(
+        "Appa 2 - ", ""
+    )
 
     return dataframe
+
 
 def get_fbk_data() -> pd.DataFrame:
     """
@@ -54,13 +54,13 @@ def get_fbk_data() -> pd.DataFrame:
         pd.DataFrame: the dataframe
     """
     dataframe = pd.read_csv(
-        os.path.join(os.path.dirname(__file__), FBK_FILE_PATH),
-        encoding='windows-1252'
+        os.path.join(os.path.dirname(__file__), FBK_FILE_PATH), encoding="windows-1252"
     )
 
     dataframe = filter_fbk_data(dataframe)
     # Drop FBK csv column
-    return dataframe.drop(["Unnamed: 0"], axis = 1)
+    return dataframe.drop(["Unnamed: 0"], axis=1)
+
 
 def get_appa_data() -> pd.DataFrame:
     """
@@ -70,9 +70,9 @@ def get_appa_data() -> pd.DataFrame:
         pd.DataFrame: the dataframe
     """
     return pd.read_csv(
-        os.path.join(os.path.dirname(__file__), APPA_FILE_PATH),
-        encoding='windows-1252'
+        os.path.join(os.path.dirname(__file__), APPA_FILE_PATH), encoding="windows-1252"
     )
+
 
 def get_prediction_data() -> pd.DataFrame:
     """
@@ -83,5 +83,5 @@ def get_prediction_data() -> pd.DataFrame:
     """
     return pd.read_csv(
         os.path.join(os.path.dirname(__file__), PREDICTION_FILE_PATH),
-        encoding='windows-1252'
+        encoding="windows-1252",
     )
